Bugfix in LabelTampering._tamper_events()
The method was originally meant to deal with LF outputs represented by an immutable data-type (ie: strings).
Since we resort to lists to implement LF outputs, I modified the LabelTampering wrapper to make a deepcopy of the original LF outputs before passing it to the tampering logic,
which is now assumed to modify its input event list in-place.