Introduce reset() method definition in Algo class
This addition is motivated by the fact that RewardMachineLearningAgent instances rely on Algo instances to provide a reset() method, despite the fact that the Algo interface does not define it. At the moment, this works because the only subclass of Algo in the codebase is QRM, which provides the method.