Implement set_save_path() for QRM and DeepQRM
Since QRM instances do not need special handling when being serialized, their implementation does effectively nothing and the save path is ignored.

On the contrary, DeepQRM need to store their Q-network parameters properly when being serialized and thus resort to this method. Specifically, the DeepQRM.set_save_path() implementation can be used as a reference for how the method is intended to be used to allow
for more complex serialization logics.