Fix bug in logging of learnt RMs to Tensorboard
Before this commit, the plots of the RMs learnt by the agents were fecthed without being sorted, leading to them being randomly sorted. This commit fixes this issue, by making sure that the plots are added to Tensorboard in the correct order: from the first learnt RM to the last one.