Fix 'logged num_steps/reward always 0' bug
Description of the issue:
When an episode is either terminated or truncated,
the RecordEpisodeStatistics wrapper updates the
entries in info["episode"] and resets its internal
counters for the episode length and reward to 0.

At the same time, the RewardMachineLearningAgent.update_agent()
method adds the current agent in the set of agents
to be interrupted when either:
- the episode is terminated OR
- the episode is truncated OR
- the agent believes the current RM state is terminal.

Then, the Trainer code checks if any agent must be
interrupted and, if so, updates the info["episode"]
entry itself, by resorting to the wrapper internal counters.

This leads to the bug: if the current agent is to
be interrupted because the episode terminated or
was truncated, the Trainer code overwrites the
info["episode"] entries by resorting to the internal
counters kept by RecordEpisodeStatistics, which were
however previously reset to 0 in the step() method.

Description of the fix:
To avoid this issue, the update_agent() method now
only adds the current agent to the list of agents
to be interrupted when the episode was not terminated
or truncated but the RM state is thought to be terminal.