Fix bug in epsilon decay mechanism
This commit fixes a subtle bug in the epsilon decay
mechanism used by the algorithm to balance between
exploration and exploitation of the learnt policy.

The issue lied in the fact that the epsilon value
to be used for action-selection was not reset to
its starting value after a call to DeepQRM.reset().
This meant that, even if the policy was discarded
and re-initialized, the epsilon value would remain
at its previous value, effectively forcing the agent to
re-learn its policy without any random exploration.

This commit fixes the issue by properly resetting
the number of training iterations so far to 0 after
a call to DeepQRM.reset()