Minor fix to RewardMachineAgent
This fix removes the call to algo.reset() in the
RewardMachineAgent.__init__() method.

The call was superfluous, as the reset() method is
intended to be used when an already trained algo
instance needs to be reset for starting training
again, eg: when a new RM is learnt. It is not intended to initialize the Algo class, as this
task is left to the Algo.__init__() method.