Improvements to DeepQRM policy reset strategies
This commit introduces new behaviour for the DeepQRM.reset() method.
Specifically, it introduces a new strategy for resetting the policy.

Before this commit, the only option was to discard
the current networks and build new ones. With this
commit, a new possibility is added: keep the same
networks and modify their parameters by adding
gaussian noise.

This is aimed at avoiding the situation where an
agent must learn everything again from scratch
after a reset() call, for instance after a new RM
is found by a learner.